backend/python/requirements.txt
flask
torch
transformers
pillow
flask-cors
gunicorn>=20.1.0
waitress>=2.1.0
PyJWT